+++ src/container/profile/index.js
@@ -1,0 +1,1 @@
+export {default as ProfileCard} from './ProfileCard'
