+++ src/container/test/index.js
@@ -1,0 +1,1 @@
+export {default as TestCard} from './TestCard'
