+++ src/template/pages/TestStart.jsx
@@ -8,48 +8,8 @@ import { Button, Card, Container, Row, Col } from 'reactstrap';
 import DemoNavbar from 'components/Navbars/DemoNavbar.js';
 import SimpleFooter from 'components/Footers/SimpleFooter.js';
 
-const TestStart = () => {
-    const [testnum, setTestnum] = useState(1)
-    const [tests, setTests] = useState(null);
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState(null);
-    const [testgen, setTestgen] = useState(1)
-
-    const handleClick = () => {
-      if (testnum < 20) {
-        setTestgen(testgen + 1)
-        setTestnum(testnum + 1)
-      } 
-      else {
-        alert('테스트가 종료되었습니다.')
-      } 
-  }
-
-    useEffect(() => {
-      const fetchTests = async () => {
-        try {
-          // 요청이 시작 할 때에는 error 와 tests 를 초기화하고
-          setError(null);
-          setTests(null);
-          // loading 상태를 true 로 바꿉니다.
-          setLoading(true);
-          const response = await axios.get(
-            'http://127.0.0.1:8080/api/legacies'
-          );
-          setTests(response.data);
-          console.log(response.data) // 데이터는 response.data 안에 들어있습니다.
-        } catch (e) {
-          setError(e);
-        }
-        setLoading(false);
-      };
-  
-      fetchTests();
-    }, []);
-    if (loading) return <div>로딩중..</div>;
-    if (error) return <div>에러가 발생했습니다</div>;
-    if (!tests) return null;
-
+const TestStart = ({children}) => {
+    
     return <>
         <DemoNavbar />
         <main className="profile-page">
@@ -75,87 +35,7 @@ const TestStart = () => {
             </div>
           </section>
           <section className="section">
-            <Container>
-              <Card className="card-profile shadow mt--300">
-                <div className="px-4">
-                  <div className="text-center mt-5">
-                    <h3>
-                      {testnum}번 문제{' '}
-                    </h3>
-                    <div className="h6 font-weight-300">
-                      <i className="ni location_pin mr-2" />
-                    </div>
-                    <div className="h6 mt-4">
-                      <i className="ni business_briefcase-24 mr-2" />
-                        {tests.[testgen].question}
-                    </div>
-                  </div>
-                  <div className="mt-5 py-5 text-center">
-                    <Row className="justify-content-center">
-                      <Col lg="9">
-                        <div className="mb-3">
-              {/* <small className="text-uppercase font-weight-bold">정답</small> */}
-            </div>
-            <div className="row-grid">
-            <Row className="justify-content-around">
-            <div className="custom-control custom-radio mb-3">
-              <input
-                className="custom-control-input"
-                id="customRadio1"
-                name="custom-radio-1"
-                type="radio"
-              />
-              <label className="custom-control-label" htmlFor="customRadio1">
-                <span>{tests.[testgen].[1]}</span>
-              </label>
-            </div>
-            <div className="custom-control custom-radio mb-3">
-              <input
-                className="custom-control-input"
-                defaultChecked
-                id="customRadio2"
-                name="custom-radio-1"
-                type="radio"
-              />
-              <label className="custom-control-label" htmlFor="customRadio2">
-                <span>{tests.[testgen].[2]}</span>
-              </label>
-            </div>
-            </Row>
-            <Row className="justify-content-around">
-            <div className="custom-control custom-radio mb-3">
-              <input
-                className="custom-control-input"
-                defaultChecked
-                id="customRadio3"
-                name="custom-radio-1"
-                type="radio"
-              />
-              <label className="custom-control-label" htmlFor="customRadio3">
-                <span>{tests.[testgen].[3]}</span>
-              </label>
-            </div>
-            <div className="custom-control custom-radio mb-3">
-              <input
-                className="custom-control-input"
-                defaultChecked
-                id="customRadio4"
-                name="custom-radio-1"
-                type="radio"
-              />
-              <label className="custom-control-label" htmlFor="customRadio4">
-                <span>{tests.[testgen].[4]}</span>
-              </label>
-            </div>
-            </Row>
-            </div>
-            <button className="float-center btn btn-default btn-lg mt-3" onClick={handleClick}>다음 문제</button>
-                      </Col>
-                    </Row>
-                  </div>
-                </div>
-              </Card>
-            </Container>
+            {children}
           </section>
         </main>
         <SimpleFooter />
